@@ -4,22 +4,22 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: 3.0.5
+  sdk: ">=2.19.0 <3.0.0"
+  flutter: ">=3.7.0"
 
 dependencies:
   bloc: ^8.1.0
   equatable: ^2.0.5
-  flame: ^1.3.0
-  flame_audio: ^1.3.1
-  flame_behaviors: ^0.1.1
-  flame_tiled: ^1.7.2
+  flame: ^1.6.0
+  flame_audio: ^1.3.5
+  flame_behaviors: ^0.2.0
+  flame_tiled: ^1.9.1
   flutter:
     sdk: flutter
   flutter_bloc: ^8.1.1
   flutter_localizations:
     sdk: flutter
-  freezed_annotation: ^2.1.0
+  freezed_annotation: ^2.2.0
   get_it: ^7.2.0
   intl: ^0.17.0
   sharp_toggle_switch:
@@ -27,13 +27,13 @@ dependencies:
 
 dev_dependencies:
   bloc_test: ^9.1.0
-  build_runner: ^2.2.0
-  flame_test: ^1.7.0
+  build_runner: ^2.3.3
+  flame_test: ^1.9.2
   flutter_test:
     sdk: flutter
   freezed: ^2.1.0+1
   mocktail: ^0.3.0
-  very_good_analysis: ^3.0.1
+  very_good_analysis: ^3.1.0
 
 flutter:
   uses-material-design: true
