@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 import 'package:one_dungeon/constants/constants.dart';
@@ -6,7 +8,7 @@ class ElevatorSprite extends SpriteComponent {
   ElevatorSprite({required super.size});
 
   @override
-  Future<void>? onLoad() {
+  FutureOr<void>? onLoad() {
     sprite = Sprite(Flame.images.fromCache(GameAssets.kElevatorImage));
     return super.onLoad();
   }
